Module sheave_server::handlers

Functions

handle_connect

Handles a connect command as a server.

handle_create_stream

Handles a createStream command as a server.

handle_fc_publish

Handles a FCPublish command as a server.

handle_first_handshake

Handles a handshake chunk of the first step as a server. This step performs:

handle_publish

Handles a publish command as a server.

handle_release_stream

Handles a releaseStream command as a server.

handle_second_handshake

Handles a handshake chunk of the second step as a server. This step performs: